+++ lib/modules/security/security_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: security
-// Updated at: 2026-05-05 21:32:51
-// Commit: test(security): implement widget integration tests for security
+// Updated at: 2026-05-05 22:48:28
+// Commit: chore(security): update dependencies and cleanup unused imports in security
 
 class securityComponentState {
-  final String commitRef = '910';
-  final String lastMessage = 'test(security): implement widget integration tests for security';
-  final DateTime timestamp = DateTime.parse('2026-05-05T21:32:51+05:00');
+  final String commitRef = '917';
+  final String lastMessage = 'chore(security): update dependencies and cleanup unused imports in security';
+  final DateTime timestamp = DateTime.parse('2026-05-05T22:48:28+05:00');
 }